requirements.txt
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
EbookLib==0.18
evdev==1.6.1
executing==1.2.0
google-api-core==2.11.0
google-api-python-client==2.48.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
Google-Images-Search==1.4.6
googleapis-common-protos==1.58.0
gTTS==2.3.1
httplib2==0.21.0
idna==3.4
ipdb==0.13.11
ipython==8.10.0
jedi==0.18.2
lxml==4.9.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
protobuf==4.22.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfiglet==0.8.post1
Pygments==2.14.0
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
python-decouple==3.7
python-resize-image==1.1.20
python-xlib==0.33
requests==2.28.2
rsa==4.9
six==1.16.0
soupsieve==2.4
stack-data==0.6.2
termcolor==1.1.0
tomli==2.0.1
traitlets==5.9.0
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
